Rename package and docs to Chronogen
