+++ docs/cli.md
@@ -1,13 +1,13 @@
 # Command Line Usage
 
-After installing the package, the `date-generator` command becomes available. The CLI mirrors the Python API and exposes parameters that are useful when building password wordlists or quick test fixtures.
+After installing the package, the `chronogen` command becomes available. The CLI mirrors the Python API and exposes parameters that are useful when building password wordlists or quick test fixtures.
 
 ## Basic usage
 
 Generate ISO formatted dates between 2000 and 2001:
 
 ```bash
-date-generator -s 2000 -e 2001 -f YYYYMMDD
+chronogen -s 2000 -e 2001 -f YYYYMMDD
 ```
 
 Output includes the end year and prints directly to `stdout`.
@@ -15,7 +15,7 @@ Output includes the end year and prints directly to `stdout`.
 ## Customising separators and casing
 
 ```bash
-date-generator -s 1999 -e 1999 -f MMDDYYYY -S "/" --case upper
+chronogen -s 1999 -e 1999 -f MMDDYYYY -S "/" --case upper
 ```
 
 This prints `MM/DD/YYYY` in uppercase, which is particularly handy when working with systems that normalise user input.
@@ -25,7 +25,7 @@ This prints `MM/DD/YYYY` in uppercase, which is particularly handy when working 
 When testing a seasonal promotion or a leap-day edge case you might not want every day of the year:
 
 ```bash
-date-generator --start 2024 --end 2024 --months 2 --days 29
+chronogen --start 2024 --end 2024 --months 2 --days 29
 ```
 
 Only February 29th is emitted.
@@ -37,7 +37,7 @@ Multiple values can be passed for both `--months` and `--days`.
 Append context or reverse chronological order to target the latest dates first:
 
 ```bash
-date-generator -s 2018 -e 2024 -f DDMMYY -S . -P corp- --suffix "!" -r
+chronogen -s 2018 -e 2024 -f DDMMYY -S . -P corp- --suffix "!" -r
 ```
 
 ## Working with custom patterns
@@ -45,7 +45,7 @@ date-generator -s 2018 -e 2024 -f DDMMYY -S . -P corp- --suffix "!" -r
 The `--pattern` option accepts any valid `strftime` string and overrides the format template and separator configuration:
 
 ```bash
-date-generator --start 1990 --end 1990 --pattern "%d%b%Y" --case lower
+chronogen --start 1990 --end 1990 --pattern "%d%b%Y" --case lower
 ```
 
 Result: `01jan1990`, `02jan1990`, etc.
@@ -55,7 +55,7 @@ Result: `01jan1990`, `02jan1990`, etc.
 Use `--output` (or `-o`) to write values to disk and choose a newline when targeting Windows tools:
 
 ```bash
-date-generator -s 2000 -e 2005 -f YYYYMMDD -o wordlists/dates.txt -n "\r\n"
+chronogen -s 2000 -e 2005 -f YYYYMMDD -o wordlists/dates.txt -n "\r\n"
 ```
 
 ## Discovering format templates
@@ -63,6 +63,6 @@ date-generator -s 2000 -e 2005 -f YYYYMMDD -o wordlists/dates.txt -n "\r\n"
 List suggested format templates along with examples:
 
 ```bash
-date-generator --list-formats
+chronogen --list-formats
 ```
-These strings are made up of contiguous `Y`, `M`, and `D` blocks. Use `YY` for short years and `YYYY` for full years. Each component is optional, so formats like `MM`, `DD`, or `MMDD` are valid. See `date-generator --help` for the full list of options and short flag aliases.
+These strings are made up of contiguous `Y`, `M`, and `D` blocks. Use `YY` for short years and `YYYY` for full years. Each component is optional, so formats like `MM`, `DD`, or `MMDD` are valid. See `chronogen --help` for the full list of options and short flag aliases.
