+++ docs/library.md
@@ -1,11 +1,11 @@
 # Library Usage
 
-The Python API is built around the :class:`date_generator.DateGenerator` class and a small configuration dataclass.  This section covers common scenarios and advanced options.
+The Python API is built around the :class:`chronogen.DateGenerator` class and a small configuration dataclass. This section covers common scenarios and advanced options.
 
 ## Generating values lazily
 
 ```python
-from date_generator import DateGenerator
+from chronogen import DateGenerator
 
 generator = DateGenerator(start_year=2020, end_year=2021, format="YYYYMMDD", separator="-")
 for value in generator.generate():
@@ -16,10 +16,10 @@ Iteration is lazy, making it safe to stream large ranges without exhausting memo
 
 ## Working with configuration objects
 
-`DateGenerator` accepts either keyword arguments or a :class:`date_generator.DateGeneratorConfig` instance. The latter is helpful when building reusable format templates:
+`DateGenerator` accepts either keyword arguments or a :class:`chronogen.DateGeneratorConfig` instance. The latter is helpful when building reusable format templates:
 
 ```python
-from date_generator import DateGenerator, DateGeneratorConfig
+from chronogen import DateGenerator, DateGeneratorConfig
 
 config = DateGeneratorConfig(
     start_year=1980,
@@ -43,7 +43,7 @@ Use the `write` method to save outputs to disk:
 
 ```python
 from pathlib import Path
-from date_generator import DateGenerator
+from chronogen import DateGenerator
 
 DateGenerator(start_year=2000, end_year=2000, format="YYYYMMDD").write(Path("wordlists/2000.txt"))
 ```
@@ -55,7 +55,7 @@ The parent directory is created automatically if required.
 To unlock month names or other locale-aware tokens, provide `custom_pattern`:
 
 ```python
-from date_generator import generate_dates
+from chronogen import generate_dates
 
 dates = generate_dates(
     start_year=2023,
@@ -73,7 +73,7 @@ When `custom_pattern` is set, the `format` and `separator` options are ignored.
 Restrict the generator to specific months or days. Values exceeding the number of days in the month are discarded automatically, so you can safely include 31 even when February is present.
 
 ```python
-from date_generator import generate_dates
+from chronogen import generate_dates
 
 dates = generate_dates(
     start_year=2022,
@@ -85,4 +85,4 @@ dates = generate_dates(
 
 ## Handling invalid input
 
-Most validation errors raise :class:`date_generator.core.DateGeneratorError`. Catch this exception to provide user-friendly feedback in downstream tools.
+Most validation errors raise :class:`chronogen.core.DateGeneratorError`. Catch this exception to provide user-friendly feedback in downstream tools.
